requirements.txt
coveralls==2.2.0
coverage==5.3.1
djangorestframework==3.12.2
Django==2.2.17
django-polymorphic==3.0.0
django-rest-polymorphic==0.1.9
django-cors-headers==3.6.0